spaceship-app/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3